Move NotficationCenter to `center.py` and made helper method available as `opengever.activity.notification_center`.
