+++ opengever/activity/__init__.py
@@ -1,16 +1,14 @@
-from opengever.activity.interfaces import IActivitySettings
-from opengever.activity.models.activity import Activity
-from opengever.activity.models.notification import Notification
-from opengever.activity.models.resource import Resource
-from opengever.activity.models.watcher import Watcher
-from plone.registry.interfaces import IRegistry
-from zope.component import getUtility
+from opengever.activity.center import DisabledNotificationCenter
+from opengever.activity.center import PloneNotificationCenter
+from opengever.activity.utils import is_activity_feature_enabled
 from zope.i18nmessageid import MessageFactory
 
 
 _ = MessageFactory("opengever.activity")
 
 
-def is_activity_feature_enabled():
-    registry = getUtility(IRegistry)
-    return registry.forInterface(IActivitySettings).is_feature_enabled
+def notification_center():
+    if not is_activity_feature_enabled():
+        return DisabledNotificationCenter()
+
+    return PloneNotificationCenter()
