+++ opengever/activity/models/__init__.py
@@ -1,0 +1,4 @@
+from opengever.activity.models.activity import Activity
+from opengever.activity.models.notification import Notification
+from opengever.activity.models.resource import Resource
+from opengever.activity.models.watcher import Watcher
