+++ opengever/activity/tests/test_handlers.py
@@ -1,7 +1,7 @@
 from ftw.builder import Builder
 from ftw.builder import create
-from opengever.activity import Activity
 from opengever.activity.events import NotificationEvent
+from opengever.activity.models import Activity
 from opengever.core.testing import OPENGEVER_FUNCTIONAL_ACTIVITY_LAYER
 from opengever.testing import FunctionalTestCase
 from plone import api
