+++ opengever/activity/tests/test_listing.py
@@ -1,7 +1,7 @@
 from ftw.builder import Builder
 from ftw.builder import create
 from ftw.testbrowser import browsing
-from opengever.activity.notification_center import NotificationCenter
+from opengever.activity.center import NotificationCenter
 from opengever.base.oguid import Oguid
 from opengever.testing import FunctionalTestCase
 from plone.app.testing import TEST_USER_ID
