+++ opengever/activity/tests/test_resolve.py
@@ -1,8 +1,8 @@
 from ftw.builder import Builder
 from ftw.builder import create
 from ftw.testbrowser import browsing
+from opengever.activity import notification_center
 from opengever.activity.browser.resolve import ResolveNotificationView
-from opengever.activity.utils import notification_center
 from opengever.base.oguid import Oguid
 from opengever.testing import FunctionalTestCase
 from plone.app.testing import TEST_USER_ID
