+++ opengever/task/tests/test_activities.py
@@ -1,8 +1,8 @@
 from ftw.builder import Builder
 from ftw.builder import create
 from ftw.testbrowser import browsing
-from opengever.activity import Activity
-from opengever.activity.utils import notification_center
+from opengever.activity import notification_center
+from opengever.activity.models import Activity
 from opengever.core.testing import OPENGEVER_FUNCTIONAL_ACTIVITY_LAYER
 from opengever.testing import FunctionalTestCase
 from plone.app.testing import TEST_USER_ID
